test(vue-query/useQuery): move suspense regression tests into 'useQuery.test.ts'
