+++ packages/vue-query/src/__tests__/useQuery.test.ts
@@ -6,10 +6,11 @@ import {
   reactive,
   ref,
 } from 'vue-demi'
-import { QueryObserver } from '@tanstack/query-core'
+import { QueryObserver, experimental_streamedQuery } from '@tanstack/query-core'
 import { queryKey, sleep } from '@tanstack/query-test-utils'
 import { useQuery } from '../useQuery'
 import { useBaseQuery } from '../useBaseQuery'
+import { useQueryClient } from '../useQueryClient'
 import type { Mock, MockedFunction } from 'vitest'
 
 vi.mock('../useQueryClient')
@@ -614,6 +615,50 @@ describe('useQuery', () => {
           state: expect.objectContaining({ status: 'error' }),
         }),
       )
+    })
+
+    it('should release suspense when setQueryData is called while fetch is in-flight', async () => {
+      const key = queryKey()
+
+      const query = useQuery({
+        queryKey: key,
+        queryFn: () => sleep(10000).then(() => 'fetched'),
+      })
+
+      const suspensePromise = query.suspense()
+
+      const queryClient = useQueryClient()
+      queryClient.setQueryData(key, 'manual data')
+
+      await vi.advanceTimersByTimeAsync(0)
+
+      const result = await suspensePromise
+      expect(result.data).toBe('manual data')
+    })
+
+    it('should release suspense when streamedQuery receives first chunk', async () => {
+      const key = queryKey()
+
+      async function* numberGenerator() {
+        await sleep(10)
+        yield 'chunk1'
+        await sleep(10)
+        yield 'chunk2'
+      }
+
+      const query = useQuery({
+        queryKey: key,
+        queryFn: experimental_streamedQuery({
+          streamFn: () => numberGenerator(),
+        }),
+      })
+
+      const suspensePromise = query.suspense()
+
+      await vi.advanceTimersByTimeAsync(10)
+
+      const result = await suspensePromise
+      expect(result.data).toStrictEqual(['chunk1'])
     })
   })
 })
